refactor: GameController에 작성되어 있던 GameService 로직과 OutputView 관련 로직을 분리
