+++ src/main/java/lotto/controller/GameController.java
@@ -1,15 +1,15 @@
 package lotto.controller;
 
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 import lotto.domain.Lotto;
+import lotto.service.GameService;
 import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class GameController {
     InputView inputView = new InputView();
+    OutputView outputView = new OutputView();
+    GameService gameService = new GameService();
 
     //전체 게임 여기서 컨트롤
     public void gameStart() {
@@ -18,12 +18,12 @@ public class GameController {
         int lotto_count = price / 1000;
 
         //2) 구매 가능한 로또 개수만큼 로또 자동 발급 기능
-        List<Lotto> lotto_list = generateLottoList(lotto_count);
+        List<Lotto> lotto_list = gameService.generateLottoList(lotto_count);
 
         //3) 발급한 로또 번호 출력 기능
-        printLottoCount(lotto_count);
+        outputView.printLottoCount(lotto_count);
         for (int i = 0; i < lotto_count; i++) {
-            printCurrentLottoList(lotto_list.get(i));
+            outputView.printCurrentLottoList(lotto_list.get(i));
         }
         //4) 당첨 번호 입력 기능
         List<Integer> win_list = inputView.readWinningNumber();
@@ -35,49 +35,6 @@ public class GameController {
 
         //7) 수익률 계산 기능
 
-    }
-
-    /**
-     * 2. 구매 가능한 로또 개수만큼 로또 자동 발급기능
-     */
-    public List<Lotto> generateLottoList(int lottoCount) {
-        List<Lotto> list = new ArrayList<>();
-        for (int i = 0; i < lottoCount; i++) {
-            //랜덤 값 자동 생성 및 세팅
-            list.add(new Lotto(getRandomNumList()));
-        }
-        return list;
-    }
-
-    static private List<Integer> getRandomNumList() {
-        List<Integer> list = Randoms.pickUniqueNumbersInRange(1, 45, 6);
-        sortRandomNumList(list); //정렬 시키기
-        return list;
-    }
-
-    static private void sortRandomNumList(List<Integer> list) {
-        Collections.sort(list);
-    }
-
-
-    /**
-     * 3. 발급한 로또 번호 출력 기능
-     */
-    public void printLottoCount(int lottoCount) {
-        System.out.println();
-        System.out.println(lottoCount + "개를 구매했습니다.");
-    }
-
-    public void printCurrentLottoList(Lotto lotto) {
-        System.out.print("[");
-        List<Integer> lotto_num_list = lotto.getNumbers();
-        //-> 문자열로 변환해서
-        List<String> lotto_str_num = lotto_num_list.stream()
-                .map(String::valueOf)
-                .collect(Collectors.toList());
-
-        System.out.print(String.join(", ", lotto_str_num));
-        System.out.println("]");
     }
